Preserve ILTrim deconstruction coverage
Move the two unsupported ILTrim deconstruction scenarios into standalone shared test cases so ConstructedTypesDataFlow continues to run. Keep strict analyzer checks for the moved warnings.



Copilot-Session: b4dc4554-6607-4109-ab22-56a5815b294d
