+++ src/coreclr/tools/ILTrim.Tests/ILTrimExpectedFailures.txt
@@ -82,8 +82,9 @@ DataFlow.CompilerGeneratedTypes
 DataFlow.CompilerGeneratedTypesNet90
 DataFlow.CompilerGeneratedTypesRelease
 DataFlow.CompilerGeneratedTypesReleaseNet90
-DataFlow.ConstructedTypesDataFlow
 DataFlow.ConstructorDataFlow
+DataFlow.DeconstructFieldTarget
+DataFlow.DeconstructUserDefinedConversion
 DataFlow.DependencyInjectionPattern
 DataFlow.DynamicDependencyDataflow
 DataFlow.ExponentialDataFlow
