+++ src/tools/illink/test/ILLink.RoslynAnalyzer.Tests/DataFlowTests.cs
@@ -149,6 +149,18 @@ namespace ILLink.RoslynAnalyzer.Tests
         }
 
         [Fact]
+        public Task DeconstructFieldTarget()
+        {
+            return RunTest();
+        }
+
+        [Fact]
+        public Task DeconstructUserDefinedConversion()
+        {
+            return RunTest();
+        }
+
+        [Fact]
         public Task DependencyInjectionPattern()
         {
             return RunTest();
